* Add a new object "ConsoleInfo" for control all info pannels on the project. * Add clock on the Alarm sound page.
